Make arrays use concrete types, not aliases.
Currently when you create an array type using something like:
```yaml
components:
  schemas:
    my_item:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
    example:
      type: array
      items:
        $ref: '#/components/schemas/my_item'
```

the generator produces a type alias for the array:
```go
// Example defines model for example.
type Example = []MyItem
```

This causes problems, particularly for defining methods. This patch
ensures that when arrays are generated, they use a concrete type and not
a type alias. For example:
```go
// Example defines model for example.
type Example []MyItem
```